Translate unknown tags in YAML
A lot of documents have tags (ex: cloudformation) which currently cause
yq to bomb out.  This translates the tags into objects with a single key
 equal to '__yq_TAG' (which hopefully keeps it unique) with the key's value
 being the tagged value.

Pyyaml doesn't expose a way to do this nicely - it looks like it expects
you to process the whole subtree, so I had to copy a chunk of the
original code from the pyyaml constructor.py source.  I omitted some
code that was there to keep the pr small, but I think it was for
features not being used in yq.