Name, parse, and pass positional-consuming jq args. Fixes #16
When a jq command line argument consumes positionals that follow it,
the yq parser must know how many positionals to pass through to it in
order to avoid getting confused about which trailing positionals are
pattern and file options.

This unfortunately means any new positional-consuming optionals that
jq introduces will need to be tracked down and added here.